requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==18.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
dill==0.3.4
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
ipapy==0.0.9.0
jamotools==0.1.10
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
networkx==2.6.2
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.1
pdflatex==0.1.3
Pillow==8.3.1
promise==2.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
regex==2021.8.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.15.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.5.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
threadpoolctl==2.2.0
tifffile==2021.8.8
torch==1.9.0
torchtext==0.10.0
tqdm==4.61.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
